[Bugfix] vecel 에러 문제 해결 #50,#51 -  getStaticPaths와 getStaticProps삭제 - /mypage에서 useEffect 사용하여 데이터 get 및 props 전달
